pyproject.toml
[tool.poetry]
name = "contact-negative-converter"
version = "0.1.0"
description = "A FastAPI web server to convert images to contact negatives using .acv curve adjustments."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
pillow = "^9.0.1"
numpy = "^1.22.3"
python-multipart = "^0.0.5"
opencv-python = "^4.7.0.72"
opencv-python-headless = "^4.7.0.72"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
